# ehr-backend-service/hosted-openapi.json
{"openapi":"3.0.1","info":{"title":"ehr_backend","contact":{},"version":"v1.0"},"servers":[{"url":"https://e8f34e4e-e9bb-4799-b725-7173d271fa62-prod.e1-us-east-azure.bijiraapis.dev/lahiru/ehrbackend/v1.0"}],"paths":{"/debug/data":{"get":{"operationId":"getDebugData","responses":{"200":{"description":"Ok","content":{"application/json":{"schema":{"oneOf":[{"type":"object"},{"$ref":"#/components/schemas/ErrorResponse"}]}}}}},"security":[{"default":[]}],"x-throttling-tier":"Unlimited","x-throttling-limit":{"requestCount":-1,"unit":"MINUTE"},"x-auth-type":"Application & Application User"}},"/orders/medication":{"post":{"operationId":"postOrdersMedication","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MedicationOrder"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/schemas/OrderResponse"},{"$ref":"#/components/schemas/ErrorResponse"}]}}}},"400":{"description":"BadRequest","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorPayload"}}}}},"security":[{"default":[]}],"x-throttling-tier":"Unlimited","x-throttling-limit":{"requestCount":-1,"unit":"MINUTE"},"x-auth-type":"Application & Application User"}},"/patients/{id}/labs":{"get":{"operationId":"getPatientsIdLabs","parameters":[{"name":"id","in":"path","required":true,"style":"simple","explode":false,"schema":{"type":"string"}},{"name":"names","in":"query","required":false,"style":"form","explode":true,"schema":{"type":"string"}},{"name":"last_n","in":"query","required":false,"style":"form","explode":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Ok","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/schemas/LabsResponse"},{"$ref":"#/components/schemas/ErrorResponse"}]}}}},"400":{"description":"BadRequest","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorPayload"}}}},"404":{"description":"NotFound"}},"security":[{"default":[]}],"x-throttling-tier":"Unlimited","x-throttling-limit":{"requestCount":-1,"unit":"MINUTE"},"x-auth-type":"Application & Application User"}},"/patients/{id}/summary":{"get":{"operationId":"getPatientsIdSummary","parameters":[{"name":"id","in":"path","required":true,"style":"simple","explode":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"Ok","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/schemas/PatientSummary"},{"$ref":"#/components/schemas/ErrorResponse"}]}}}},"400":{"description":"BadRequest","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorPayload"}}}},"404":{"description":"NotFound"}},"security":[{"default":[]}],"x-throttling-tier":"Unlimited","x-throttling-limit":{"requestCount":-1,"unit":"MINUTE"},"x-auth-type":"Application & Application User"}}},"components":{"schemas":{"Demographics":{"type":"object","properties":{"address":{"type":"string"},"dateOfBirth":{"type":"string"},"email":{"type":"string"},"firstName":{"type":"string"},"gender":{"type":"string"},"lastName":{"type":"string"},"patientId":{"type":"string"},"phone":{"type":"string"}},"required":["dateOfBirth","firstName","gender","lastName","patientId"]},"ErrorPayload":{"type":"object","properties":{"message":{"type":"string"},"method":{"type":"string"},"path":{"type":"string"},"reason":{"type":"string"},"status":{"type":"integer","format":"int64"},"timestamp":{"type":"string"}},"required":["message","method","path","reason","status","timestamp"]},"ErrorResponse":{"type":"object","properties":{"error":{"type":"string"},"message":{"type":"string"}},"required":["error","message"]},"LabResult":{"type":"object","properties":{"labId":{"type":"string"},"name":{"type":"string"},"recordDate":{"type":"string"},"referenceRange":{"type":"string"},"status":{"type":"string"},"unit":{"type":"string"},"value":{"type":"number","format":"double"}},"required":["labId","name","recordDate","referenceRange","status","unit","value"]},"LabsResponse":{"type":"object","properties":{"labs":{"type":"array","items":{"$ref":"#/components/schemas/LabResult"}},"patientId":{"type":"string"}},"required":["labs","patientId"]},"Medication":{"type":"object","properties":{"dosage":{"type":"string"},"frequency":{"type":"string"},"medicationId":{"type":"string"},"name":{"type":"string"},"prescribedDate":{"type":"string"},"status":{"type":"string"}},"required":["dosage","frequency","medicationId","name","prescribedDate","status"]},"MedicationOrder":{"type":"object","properties":{"dosage":{"type":"string"},"duration":{"type":"string"},"frequency":{"type":"string"},"medicationName":{"type":"string"},"notes":{"type":"string"},"patientId":{"type":"string"},"prescriberId":{"type":"string"}},"required":["dosage","duration","frequency","medicationName","patientId","prescriberId"]},"OrderResponse":{"type":"object","properties":{"message":{"type":"string"},"orderId":{"type":"string"},"status":{"type":"string"}},"required":["message","orderId","status"]},"PatientSummary":{"type":"object","properties":{"demographics":{"$ref":"#/components/schemas/Demographics"},"lastA1c":{"allOf":[{"$ref":"#/components/schemas/LabResult"}]},"lastEgfr":{"allOf":[{"$ref":"#/components/schemas/LabResult"}]},"medications":{"type":"array","items":{"$ref":"#/components/schemas/Medication"}},"problems":{"type":"array","items":{"$ref":"#/components/schemas/Problem"}},"vitals":{"type":"array","items":{"$ref":"#/components/schemas/Vitals"}}},"required":["demographics","lastA1c","lastEgfr","medications","problems","vitals"]},"Problem":{"type":"object","properties":{"description":{"type":"string"},"onsetDate":{"type":"string"},"problemId":{"type":"string"},"severity":{"type":"string"},"status":{"type":"string"}},"required":["description","problemId","status"]},"Vitals":{"type":"object","properties":{"bloodPressureDiastolic":{"type":"number","format":"double"},"bloodPressureSystolic":{"type":"number","format":"double"},"heartRate":{"type":"number","format":"double"},"height":{"type":"number","format":"double"},"recordDate":{"type":"string"},"temperature":{"type":"number","format":"double"},"weight":{"type":"number","format":"double"}},"required":["bloodPressureDiastolic","bloodPressureSystolic","heartRate","height","recordDate","temperature","weight"]}},"securitySchemes":{"default":{"type":"oauth2","flows":{"implicit":{"authorizationUrl":"https:///authorize","scopes":{}}}}}},"x-wso2-disable-security":"false"}