# evidence_agent_mcp.py
#!/usr/bin/env python3
"""
Evidence Agent with MCP Integration

This version of the evidence agent uses MCP (Model Context Protocol) servers
to fetch data instead of direct REST API calls.
"""

import json
import logging
import os
import re
from datetime import datetime
from typing import Any, Dict, List, Optional, TypedDict

import requests
from dotenv import load_dotenv
from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware

load_dotenv()

logger = logging.getLogger("evidence_agent_mcp")
if not logger.handlers:
    handler = logging.StreamHandler()
    formatter = logging.Formatter(
        "%(asctime)s [%(levelname)s] evidence-agent-mcp: %(message)s"
    )
    handler.setFormatter(formatter)
    logger.addHandler(handler)
logger.setLevel(logging.INFO)

def _strip_quotes(value: Optional[str]) -> Optional[str]:
    """Strip quotes from environment variable values."""
    if value is None:
        return None
    value = value.strip()
    if len(value) >= 2 and value[0] == value[-1] and value[0] in {"'", '"'}:
        return value[1:-1]
    return value

# MCP Gateway configuration
MCP_GATEWAY_CLIENT_ID = _strip_quotes(os.environ.get("MCP_GATEWAY_CLIENT_ID"))
MCP_GATEWAY_CLIENT_SECRET = _strip_quotes(os.environ.get("MCP_GATEWAY_CLIENT_SECRET"))
MCP_GATEWAY_TOKEN_ENDPOINT = _strip_quotes(os.environ.get("MCP_GATEWAY_TOKEN_ENDPOINT"))

# MCP Server URLs
EHR_MCP_URL = _strip_quotes(os.environ.get("EHR_MCP_URL"))
TRIAL_REGISTRY_MCP_URL = _strip_quotes(os.environ.get("TRIAL_REGISTRY_MCP_URL"))

# Cache for MCP access token
_mcp_access_token_cache = {"token": None, "expires_at": 0}

class MCPClient:
    """Client for interacting with MCP servers."""
    
    def __init__(self, server_url: str):
        self.server_url = server_url
        self._access_token = None
    
    def _get_access_token(self) -> Optional[str]:
        """Get cached or fetch new MCP access token."""
        if not all([MCP_GATEWAY_CLIENT_ID, MCP_GATEWAY_CLIENT_SECRET, MCP_GATEWAY_TOKEN_ENDPOINT]):
            logger.error("MCP Gateway credentials not configured")
            return None
        
        import time
        if _mcp_access_token_cache["token"] and _mcp_access_token_cache["expires_at"] > time.time() + 60:
            return _mcp_access_token_cache["token"]
        
        logger.info("Requesting new MCP access token")
        try:
            response = requests.post(
                str(MCP_GATEWAY_TOKEN_ENDPOINT),
                data={
                    "grant_type": "client_credentials",
                    "client_id": MCP_GATEWAY_CLIENT_ID,
                    "client_secret": MCP_GATEWAY_CLIENT_SECRET,
                },
                headers={"Content-Type": "application/x-www-form-urlencoded"},
                timeout=10,
            )
            response.raise_for_status()
            token_data = response.json()
            
            access_token = token_data["access_token"]
            expires_in = token_data.get("expires_in", 3600)
            
            _mcp_access_token_cache["token"] = access_token
            _mcp_access_token_cache["expires_at"] = time.time() + expires_in
            
            logger.info("✓ MCP access token obtained (expires in %d seconds)", expires_in)
            return access_token
            
        except Exception as e:
            logger.error("✗ Failed to obtain MCP access token: %s", e)
            return None
    
    def call_tool(self, tool_name: str, arguments: Dict[str, Any]) -> Optional[Dict[str, Any]]:
        """Call an MCP tool."""
        access_token = self._get_access_token()
        if not access_token:
            return None
        
        payload = {
            "jsonrpc": "2.0",
            "id": 1,
            "method": "tools/call",
            "params": {
                "name": tool_name,
                "arguments": arguments
            }
        }
        
        headers = {
            "Authorization": f"Bearer {access_token}",
            "Content-Type": "application/json",
        }
        
        try:
            logger.info("Calling MCP tool '%s' at %s", tool_name, self.server_url)
            response = requests.post(
                self.server_url,
                headers=headers,
                json=payload,
                timeout=30,
            )
            response.raise_for_status()
            result = response.json()
            
            if "error" in result:
                logger.error("MCP tool error: %s", result["error"])
                return None
            
            logger.info("✓ MCP tool '%s' executed successfully", tool_name)
            return result.get("result")
            
        except Exception as e:
            logger.error("✗ MCP tool call failed: %s", e)
            return None

# Initialize MCP clients
ehr_mcp_client = MCPClient(EHR_MCP_URL) if EHR_MCP_URL else None
trial_mcp_client = MCPClient(TRIAL_REGISTRY_MCP_URL) if TRIAL_REGISTRY_MCP_URL else None

class GeoFilter(object):
    def __init__(self, lat: float, lon: float, radius_km: float = 25.0):
        self.lat = lat
        self.lon = lon
        self.radius_km = radius_km

class EvidenceRequest(object):
    def __init__(self, age: int, diagnosis: str, egfr: float, comorbidities: Optional[List[str]] = None, geo: Optional[GeoFilter] = None):
        self.age = age
        self.diagnosis = diagnosis
        self.egfr = egfr
        self.comorbidities = comorbidities or []
        self.geo = geo

def fetch_patient_data_via_mcp(patient_id: str) -> Optional[Dict[str, Any]]:
    """Fetch patient data using EHR MCP server."""
    if not ehr_mcp_client:
        logger.error("EHR MCP client not configured")
        return None
    
    logger.info("Fetching patient data via MCP for patient_id=%s", patient_id)
    
    # Try to get patient summary
    summary_result = ehr_mcp_client.call_tool("getPatientsIdSummary", {"id": patient_id})
    if summary_result and not summary_result.get("isError", True):
        logger.info("✓ Patient summary fetched via MCP")
        return summary_result
    
    logger.warning("Failed to fetch patient summary via MCP, using fallback")
    return None

def fetch_trials_via_mcp() -> List[Dict[str, Any]]:
    """Fetch clinical trials using Trial Registry MCP server."""
    if not trial_mcp_client:
        logger.error("Trial Registry MCP client not configured")
        return []
    
    logger.info("Fetching trials via MCP")
    
    # Try to get trials list
    trials_result = trial_mcp_client.call_tool("getTrials", {})
    if trials_result and not trials_result.get("isError", True):
        logger.info("✓ Trials data fetched via MCP")
        trials_data = trials_result.get("content", [])
        if trials_data and isinstance(trials_data[0], dict) and "text" in trials_data[0]:
            try:
                return json.loads(trials_data[0]["text"])
            except json.JSONDecodeError:
                logger.error("Failed to parse trials JSON from MCP")
        return trials_data
    
    logger.warning("Failed to fetch trials via MCP, using fallback")
    return []

def generate_evidence_pack_mcp(request: EvidenceRequest) -> Dict[str, Any]:
    """Generate evidence pack using MCP services."""
    logger.info("Generating evidence pack using MCP services")
    logger.info("Request: age=%d, diagnosis=%s, egfr=%.1f", request.age, request.diagnosis, request.egfr)
    
    # Fetch trials via MCP
    trials = fetch_trials_via_mcp()
    
    # If MCP fails, use fallback synthetic data
    if not trials:
        logger.info("Using synthetic trial data as fallback")
        trials = [
            {
                "id": 1,
                "nct_id": "NCT05501234",
                "title": "SGLT2 Inhibitor Study for CKD Patients",
                "condition": "Type 2 diabetes mellitus with chronic kidney disease",
                "phase": "Phase 3",
                "status": "Recruiting",
                "eligibility_summary": "eGFR 30-60, HbA1c 7-10%",
                "site_distance_km": 15.2
            },
            {
                "id": 2,
                "nct_id": "NCT05612345",
                "title": "GLP-1 RA in Advanced CKD",
                "condition": "Chronic kidney disease stage 3-4",
                "phase": "Phase 2",
                "status": "Active",
                "eligibility_summary": "Adults with eGFR 15-60",
                "site_distance_km": 22.8
            }
        ]
    
    # Score and filter trials
    scored_trials = []
    for trial in trials:
        score = 0.0
        
        # Score based on diagnosis match
        condition = trial.get("condition", "").lower()
        if request.diagnosis.lower() in condition:
            score += 2.0
        
        # Score based on CKD relevance
        if "ckd" in condition or "kidney" in condition:
            score += 1.0
        
        # Score based on eGFR criteria
        eligibility = trial.get("eligibility_summary", "")
        if eligibility and "eGFR" in eligibility and request.egfr <= 60:
            score += 1.0
        
        # Score based on age (if relevant)
        if request.age >= 60:
            score += 0.5
        
        # Filter by location if provided
        distance = trial.get("site_distance_km")
        if request.geo and distance and distance > request.geo.radius_km:
            continue
        
        scored_trials.append({
            "id": trial.get("id"),
            "nct_id": trial.get("nct_id"),
            "title": trial.get("title"),
            "condition": trial.get("condition"),
            "phase": trial.get("phase"),
            "status": trial.get("status"),
            "site_distance_km": distance,
            "suitability": round(score, 2),
            "why_match": eligibility or f"Matches diagnosis {request.diagnosis}"
        })
    
    # Sort by suitability and take top 3
    scored_trials.sort(key=lambda x: x["suitability"], reverse=True)
    top_trials = scored_trials[:3]
    
    # Generate evidence analyses
    analyses = []
    for trial in top_trials:
        suitability = trial["suitability"]
        if suitability >= 2.5:
            grade = "high"
            benefit = "Strong evidence for efficacy in similar populations"
            risk = "Monitor renal function and drug interactions"
        elif suitability >= 1.5:
            grade = "medium"
            benefit = "Moderate evidence with potential benefits"
            risk = "Consider individual patient factors"
        else:
            grade = "low"
            benefit = "Limited evidence for this specific population"
            risk = "Uncertain benefit-risk profile"
        
        analyses.append({
            "trial_id": trial["id"],
            "trial_title": trial["title"],
            "pico_grade": grade,
            "benefit_summary": benefit,
            "risk_summary": risk,
            "overall_summary": f"{trial['title']} shows {grade} relevance for {request.diagnosis} with eGFR {request.egfr}"
        })
    
    # Build evidence pack
    evidence_pack = {
        "patient": {
            "age": request.age,
            "diagnosis": request.diagnosis,
            "egfr": request.egfr,
            "comorbidities": request.comorbidities,
            "geo": {
                "lat": request.geo.lat,
                "lon": request.geo.lon,
                "radius_km": request.geo.radius_km
            } if request.geo else None
        },
        "analyses": analyses,
        "trials": top_trials,
        "generated_at": datetime.utcnow().isoformat() + "Z",
        "mcp_enabled": True,
        "data_sources": {
            "ehr_mcp": bool(ehr_mcp_client),
            "trial_mcp": bool(trial_mcp_client)
        }
    }
    
    logger.info("✓ Evidence pack generated with %d analyses and %d trials", len(analyses), len(top_trials))
    return evidence_pack

# FastAPI app
app = FastAPI(
    title="Evidence Agent with MCP",
    description="Evidence aggregation service using Model Context Protocol",
    version="1.0.0-mcp",
)

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=False,
    allow_methods=["*"],
    allow_headers=["*"],
)

@app.post("/agents/evidence/search")
def evidence_search_mcp(payload: dict) -> dict:
    """Evidence search endpoint using MCP integration."""
    logger.info("Evidence search request received via MCP endpoint")
    
    try:
        # Parse request
        age = payload.get("age")
        diagnosis = payload.get("diagnosis")
        egfr = payload.get("egfr")
        comorbidities = payload.get("comorbidities", [])
        geo_data = payload.get("geo")
        
        if not all([age is not None, diagnosis, egfr is not None]):
            raise HTTPException(status_code=400, detail="Missing required fields: age, diagnosis, egfr")
        
        geo = None
        if geo_data:
            geo = GeoFilter(
                lat=geo_data.get("lat", 35.15),
                lon=geo_data.get("lon", -90.05),
                radius_km=geo_data.get("radius_km", 25.0)
            )
        
        request = EvidenceRequest(
            age=int(age) if age is not None else 0,
            diagnosis=str(diagnosis) if diagnosis else "",
            egfr=float(egfr) if egfr is not None else 0.0,
            comorbidities=comorbidities,
            geo=geo
        )
        
        # Generate evidence pack using MCP
        evidence_pack = generate_evidence_pack_mcp(request)
        
        return {"evidence_pack": evidence_pack}
        
    except Exception as e:
        logger.exception("Evidence search failed: %s", e)
        raise HTTPException(status_code=500, detail=str(e))

@app.get("/health")
def health_check() -> dict:
    """Health check endpoint."""
    mcp_status = {
        "ehr_mcp_configured": bool(ehr_mcp_client),
        "trial_mcp_configured": bool(trial_mcp_client),
        "mcp_credentials_configured": bool(MCP_GATEWAY_CLIENT_ID and MCP_GATEWAY_CLIENT_SECRET)
    }
    
    return {
        "status": "ok",
        "version": "1.0.0-mcp",
        "mcp_integration": mcp_status
    }

@app.get("/mcp/test")
def test_mcp_connection() -> dict:
    """Test MCP server connections."""
    results = {}
    
    if ehr_mcp_client:
        try:
            # Test EHR MCP connection
            result = ehr_mcp_client.call_tool("getDebugData", {})
            results["ehr_mcp"] = {
                "status": "connected" if result else "error",
                "url": EHR_MCP_URL
            }
        except Exception as e:
            results["ehr_mcp"] = {
                "status": "error",
                "error": str(e),
                "url": EHR_MCP_URL
            }
    else:
        results["ehr_mcp"] = {"status": "not_configured"}
    
    if trial_mcp_client:
        try:
            # Test Trial Registry MCP connection
            result = trial_mcp_client.call_tool("getTrials", {})
            results["trial_mcp"] = {
                "status": "connected" if result else "error",
                "url": TRIAL_REGISTRY_MCP_URL
            }
        except Exception as e:
            results["trial_mcp"] = {
                "status": "error",
                "error": str(e),
                "url": TRIAL_REGISTRY_MCP_URL
            }
    else:
        results["trial_mcp"] = {"status": "not_configured"}
    
    return {"mcp_connections": results}

if __name__ == "__main__":
    import uvicorn
    logger.info("Starting Evidence Agent with MCP Integration")
    uvicorn.run(app, host="0.0.0.0", port=8005)